Add Schema.from_definition
