+++ lib/graphql/schema.rb
@@ -10,7 +10,7 @@ require "graphql/schema/type_map"
 require "graphql/schema/unique_within_type"
 require "graphql/schema/validation"
 require "graphql/schema/warden"
-require "graphql/schema/build_from_ast"
+require "graphql/schema/build_from_definition"
 
 module GraphQL
   # A GraphQL schema which may be queried with {GraphQL::Query}.
@@ -45,6 +45,9 @@ module GraphQL
   #   end
   #
   class Schema
+    class InvalidDocumentError < Error; end;
+    extend BuildFromDefinition
+
     include GraphQL::Define::InstanceDefinable
     accepts_definitions \
       :query, :mutation, :subscription,
